EVM: fix overflow behaviour of addmod and mulmod
As per the Ethereum yellow paper, the intermediate results of the addmod and
mulmod instructions should not be truncated to 256-bits before performing the
modulo operation. This commit performs the truncation after the modulo.

The previous implementation resulted in incorrect results in many cases
involving large numbers, for example:

mulmod(2^255, 3, 5) = 3 (old implementation)

mulmod(2^255, 3, 5) = 4 (new implementation, correct)